refactor: move setCurrentPatient() function
